started work on questionmark events
stared working on the question mark events. worked out a system that uses mapoption struct to save all information needed (so far). saved is the text display, what happens when clicked and when to lock the option. a lot of notes were made. also have to rework the mapgen a small bit as the event is now saved in the questionmark not only if good/bad